+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [5.3.1] - 2026-02-04
+
 ### Fixed
 
 - instant scrolling to top on active `PanelSideBarItem`; scrolling smoothing to center instead.
@@ -290,12 +292,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - created package :tada:
 - added tests
 
-[unreleased]: https://github.com/neolution-ch/react-pattern-ui/compare/5.3.0...HEAD
+[unreleased]: https://github.com/neolution-ch/react-pattern-ui/compare/5.3.1...HEAD
 [2.1.1]: https://github.com/neolution-ch/react-pattern-ui/compare/2.1.0...2.1.1
 [2.1.0]: https://github.com/neolution-ch/react-pattern-ui/compare/2.0.2...2.1.0
 [2.0.2]: https://github.com/neolution-ch/react-pattern-ui/compare/2.0.1...2.0.2
 [2.0.1]: https://github.com/neolution-ch/react-pattern-ui/compare/2.0.0...2.0.1
 [2.0.0]: https://github.com/neolution-ch/react-pattern-ui/compare/45cbeb76034667019da84605082679900f506d75...2.0.0
+[5.3.1]: https://github.com/neolution-ch/react-pattern-ui/compare/5.3.0...5.3.1
 [5.3.0]: https://github.com/neolution-ch/react-pattern-ui/compare/5.2.0...5.3.0
 [5.2.0]: https://github.com/neolution-ch/react-pattern-ui/compare/5.1.0...5.2.0
 [5.1.0]: https://github.com/neolution-ch/react-pattern-ui/compare/5.0.0...5.1.0